% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/area_um_to_ecd_nm.R
\name{area_um_to_ecd_nm}
\alias{area_um_to_ecd_nm}
\title{Calculate the equivalent circular diameter cut-off in nm from the
area in microns}
\usage{
area_um_to_ecd_nm(area_um)
}
\arguments{
\item{area_um}{- the equivalent circular diameter of an object in nm.
Example: 7.854}
}
\value{
ecd in nm.
}
\description{
A convenience function to make Fiji/ImageJ particle analysis easier.
Often we want to measure the diameter of nanoparticles in nm units
but work with images with scale factors in microns/px. The 
AnalyzeParticles function wants a lower area cutoff in microns.
}
\examples{
library(particlesizeR)
print(area_um_to_ecd_nm(7.853982e-05))

}
