% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ecd_nm_to_area_um.R
\name{ecd_nm_to_area_um}
\alias{ecd_nm_to_area_um}
\title{Calculate the area cut-off in microns from the equivaent circular diameter in nm}
\usage{
ecd_nm_to_area_um(ecd_nm)
}
\arguments{
\item{ecd_nm}{- the equivalent circular diameter of an object in nm.
Example: 10.0}
}
\value{
area_um the area in microns. This Will print in exponential
notation.
}
\description{
A convenience function to make Fiji/ImageJ particle analysis easier.
Often we want to measure the diameter of nanoparticles in nm units
but work with images with scale factors in microns/px. The 
AnalyzeParticles function wants a lower area cutoff in microns.
}
\examples{
library(particlesizeR)
print(ecd_nm_to_area_um(10.0))

}
