% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gen.two.gaussian.dist.R
\name{gen.two.gaussian.dist}
\alias{gen.two.gaussian.dist}
\title{Generate a distribution from a two Gaussian mixture}
\usage{
gen.two.gaussian.dist(p1, mu1, s1, mu2, s2, start, stop, step,
  rn.seed = 42)
}
\arguments{
\item{p1}{- the fractiona probability of the first Gaussian (< 1.0)}

\item{mu1}{- the mean of the first Gaussian}

\item{s1}{- the standard deviation of the first Gaussian}

\item{mu2}{- the mean of the second Gaussian}

\item{s2}{- the standard deviation of the second Gaussian}

\item{start}{- starting value of x (e.g. diameter)}

\item{stop}{- ending value of x (e.g. diameter)}

\item{step}{- step x (e.g. diameter)}

\item{rn.seed}{- the random number seed (default 42)}
}
\value{
df - a data frame with x and frequency
}
\description{
Generate a distribution from a two Gaussian mixture
}
\examples{
library(particlesizeR)
data <- gen.two.gaussian.dist(0.5, 0.5, 1.0, 10.0, 3.0, 0, 20, 0.1)

}
