% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_circ.R
\name{ggplot_circ}
\alias{ggplot_circ}
\title{Plot the circularity distribution}
\usage{
ggplot_circ(circ, bin_width, title, xlab, ylab = "Counts",
  base_txt_pts = 12)
}
\arguments{
\item{circ}{vector - Containing the circularity}

\item{bin_width}{number - The width of the bin for the circularity histogram}

\item{title}{String - A title for the plot}

\item{xlab}{String - the label for the x axis, e.g. "diameter [nm]"}

\item{ylab}{String - the y label. Default = "Counts"}

\item{base_txt_pts}{String - the size in points. Default 12.}
}
\value{
plt
}
\description{
Make a ggplot2 plot of the traditional circularity distribution
}
\examples{

library(particlesizeR)
# simulate a distribution
vec1 <- rnorm(1000, mean=0.90, sd=.05)
ggplot_circ(vec1, 0.01, "Circularity distribution", "Circularity", "Counts")

}
