% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_ecd.R
\name{ggplot_ecd}
\alias{ggplot_ecd}
\title{Plot a histogram of Equivalent Circular Diameter data}
\usage{
ggplot_ecd(ecd, bin_width, title, xlab = "ECD [nm]", ylab = "Counts",
  base_txt_pts = 12)
}
\arguments{
\item{ecd}{vector - Contains the ECD}

\item{bin_width}{number - The width of the bin for the diameter histogram}

\item{title}{String - A title for the plot}

\item{xlab}{String - the label for the x axis, e.g. "ECD [nm]"}

\item{ylab}{String - the y label. Default = "Counts"}

\item{base_txt_pts}{String - the size in points. Default 12.}
}
\value{
plt
}
\description{
Make a ggplot2 histogram plot of the ECD
}
\examples{
library(particlesizeR)
# simulate a complicated distribution
vec1 <- rnorm(500, mean=85, sd=2)
vec2 <- rnorm(1000, mean=90, sd=2)
vec <- c(vec1, vec2)

}
