% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_ecd_boxplot.R
\name{ggplot_ecd_boxplot}
\alias{ggplot_ecd_boxplot}
\title{Plot a box plot of Equivalent Circular Diameter data}
\usage{
ggplot_ecd_boxplot(ecd, c_fixed = 0.1, title = "Box Plot",
  ylab = "ECD [nm]", base_txt_pts = 12)
}
\arguments{
\item{ecd}{vector - Contains the ECD}

\item{c_fixed}{number default: 0.1 The width parameter}

\item{title}{String - A title for the plot. Default: "Box Plot"}

\item{ylab}{String - the y axis label. Default = "ECD [nm]"}

\item{base_txt_pts}{String - the size in points. Default 12.}
}
\value{
plt
}
\description{
Make a ggplot2 boxplot of the ECD
}
\examples{
library(particlesizeR)
# simulate a distribution
vec <- rnorm(1000, mean=90, sd=2)
ggplot_ecd_boxplot(vec, 0.2, "test", "ECD [nm]", 12)

}
