% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_ecd_qq_plot.R
\name{ggplot_ecd_qq_plot}
\alias{ggplot_ecd_qq_plot}
\title{Plot a quantile-quantile of Equivalent Circular Diameter data}
\usage{
ggplot_ecd_qq_plot(ecd, c_fixed = 0.5, title = "QQ Plot",
  ylab = "ECD [nm]", base_txt_pts = 12)
}
\arguments{
\item{ecd}{vector - Contains the ECD}

\item{c_fixed}{number default: 0.1 The width parameter}

\item{title}{String - A title for the plot. Default: "QQ Plot"}

\item{ylab}{String - the y axis label. Default = "ECD [nm]"}

\item{base_txt_pts}{String - the size in points. Default 12.}
}
\value{
plt
}
\description{
Make a ggplot2 Q-Q plot of the ECD
}
\examples{
library(particlesizeR)
# simulate a distribution
ecd <- rnorm(1000, mean=90, sd=2)
ggplot_ecd_qq_plot(ecd, 0.5, "QQ Plot", "ECD [nm]", 12)

}
