% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ggpanel_plot.R
\name{make_ggpanel_plot}
\alias{make_ggpanel_plot}
\title{Make a ggplot panel plot}
\usage{
make_ggpanel_plot(ecd_plt, box_plt, ecd_qq_plt, plt_file)
}
\arguments{
\item{ecd_plt}{ggplot-plot of the ECD distribution}

\item{box_plt}{ggplot-box-plot of the ECD distribution}

\item{ecd_qq_plt}{ggplot-qq-plot of the ECD Quantile-Quantile distribution}

\item{plt_file}{string - the path to file to save, e.g. "~/my_panel_plt.png"}
}
\value{
None. Save file
}
\description{
Make a ggplot2 boxplot of the ECD
}
