% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_ecd.R
\name{ggplot_ecd}
\alias{ggplot_ecd}
\title{Plot a histogram of Equivalent Circular Diameter data}
\usage{
ggplot_ecd(ecd, bin_width, title, xlab, ylab = "Counts",
  base_txt_pts = 12)
}
\arguments{
\item{ecd}{vector - Contains the ECD}

\item{bin_width}{number - The width of the bin for the diameter histogram}

\item{title}{String - A title for the plot}

\item{xlab}{String - the label for the x axis, e.g. "diameter [nm]"}

\item{ylab}{String - the y label. Default = "Counts"}

\item{base_txt_pts}{String - the size in points. Default 12.}
}
\value{
plt
}
\description{
Make a ggplot2 histogram plot of the ECD
}
\examples{
library(particlesizeR)
vec1 <- rnorm(1000, mean=50, sd=1)
vec2 <- rnorm(1000, mean=75, sd=1.2)
vec <- vec1 + vec2
ggplot_ecd(vec, 0.2, "test", "nm")

}
